+++ src/Facades/HelloPeter.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Eugenefvdm\Api\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @method static array getUnrepliedReviews()
+ * 
+ * @see \Eugenefvdm\Api\HelloPeter
+ */
+class HelloPeter extends Facade
+{
+    /**
+     * Get the registered name of the component.
+     *
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return 'hello_peter';
+    }
+}