+++ src/Facades/Slack.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Eugenefvdm\Api\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @method static mixed sendMessage(string $message)
+ * 
+ * @see \Eugenefvdm\Api\Slack
+ */
+class Slack extends Facade
+{
+    /**
+     * Get the registered name of the component.
+     *
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return 'slack';
+    }
+}