+++ src/Facades/Telegram.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Eugenefvdm\Api\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @method static mixed sendMessage(string $message)
+ * 
+ * @see \Eugenefvdm\Api\Telegram
+ */
+class Telegram extends Facade
+{
+    /**
+     * Get the registered name of the component.
+     *
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return 'telegram';
+    }
+}