+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 21 March 2025
 
+- Updated README.md file to specify new examples for testing
+- Converted the lot to a service provider with facades access
 - Changed namespace from eugenevdm to eugenefvdm
 - Added this change log
 - Added an example to get owner email from ZA Domains
-